test-requirements.txt
thrift
logzero
hbase-thrift
sqlalchemy
newspaper3k
grequests
scrapy
scrapyd
scrapyd-client
scrapy-redis
pyquery
lxml
